refactor code and bring script can run
